+++ test/project_types/extension/features/argo_serve_test.rb
@@ -9,14 +9,14 @@ module Extension
 
       def setup
         ShopifyCli::ProjectType.load_type(:extension)
-        stub_package_manager
         super
       end
 
       def test_argo_serve_defers_to_js_system_when_shopifolk_check_is_false
+        stub_package_manager
         argo_serve = Features::ArgoServe.new(
           context: @context,
-          argo_runtime: argo_runtime,
+          argo_runtime: admin_ui_extension_runtime,
           specification_handler: specification_handler,
           js_system: fake_js_system
         )
@@ -26,9 +26,10 @@ module Extension
       end
 
       def test_argo_serve_abort_when_server_start_failed
+        stub_package_manager
         argo_serve = Features::ArgoServe.new(
           context: @context,
-          argo_runtime: argo_runtime,
+          argo_runtime: admin_ui_extension_runtime,
           specification_handler: specification_handler,
           js_system: fake_js_system(success: false)
         )
@@ -43,14 +44,58 @@ module Extension
         )
       end
 
-      private
+      def test_forwards_resource_url
+        ShopifyCli::Tasks::EnsureDevStore.stubs(:call)
+        ShopifyCli::Tasks::EnsureEnv.stubs(:call)
+        ExtensionProject.stubs(:update_env_file)
+        ExtensionProject.any_instance.expects(:resource_url).returns("/test").at_least_once
 
-      def argo_runtime
-        Features::ArgoRuntime.find(cli_package: cli, identifier: "PRODUCT_SUBSCRIPTION")
+        js_system = mock
+        js_system.expects(:call)
+          .with do |_, config|
+            assert_includes config.fetch(:yarn), "--resourceUrl=/test"
+            assert_includes config.fetch(:npm), "--resourceUrl=/test"
+          end
+          .returns(true)
+
+        argo_serve = Features::ArgoServe.new(
+          context: @context,
+          argo_runtime: checkout_ui_extension_runtime,
+          specification_handler: specification_handler,
+          js_system: js_system
+        )
+
+        argo_serve.call
       end
 
-      def cli
-        Models::NpmPackage.new(name: "@shopify/admin-ui-extensions-run", version: "0.11.0")
+      def test_builds_resource_url_if_necessary
+        ShopifyCli::Tasks::EnsureDevStore.stubs(:call)
+        ShopifyCli::Tasks::EnsureEnv.stubs(:call)
+        ExtensionProject.expects(:update_env_file).with(
+          has_entries(context: anything, resource_url: "/generated")
+        )
+
+        argo_serve = Features::ArgoServe.new(
+          context: @context,
+          argo_runtime: checkout_ui_extension_runtime,
+          specification_handler: specification_handler.tap do |handler|
+            handler.expects(:supplies_resource_url?).returns(true)
+            handler.expects(:build_resource_url).returns("/generated")
+          end,
+          js_system: fake_js_system
+        )
+
+        argo_serve.call
+      end
+
+      private
+
+      def admin_ui_extension_runtime
+        Features::Runtimes::Admin.new
+      end
+
+      def checkout_ui_extension_runtime
+        Features::Runtimes::CheckoutUiExtension.new
       end
 
       def specification_handler
@@ -59,6 +104,10 @@ module Extension
 
       def fake_js_system(success: true)
         proc { success }
+      end
+
+      def stub_ensure_env_check
+        ShopifyCli::Tasks::EnsureEnv.stubs(:call)
       end
 
       def stub_package_manager